Rename RealTimeProviders to CSVDocuments
so we can add more stuff into it after
