+++ apps/transport/lib/transport/application.ex
@@ -5,7 +5,7 @@ defmodule Transport.Application do
   """
 
   use Application
-  alias Transport.{ImportDataWorker, RealTimeProviders, SearchCommunes}
+  alias Transport.{CSVDocuments, ImportDataWorker, SearchCommunes}
   alias TransportWeb.Endpoint
   import Supervisor.Spec, only: [supervisor: 2]
 
@@ -14,7 +14,7 @@ defmodule Transport.Application do
     children = [
       supervisor(TransportWeb.Endpoint, []),
       supervisor(ImportDataWorker, []),
-      RealTimeProviders,
+      CSVDocuments,
       SearchCommunes
     ]
     |> add_scheduler()
